requirements_legacy.txt
vaderSentiment
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.9
amqp==2.3.2
appdirs==1.4.3
arviz==0.7.0
asgiref==3.2.7
asn1crypto==1.3.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.6.0
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
billiard==3.5.0.4
bleach==3.1.0
bs4==0.0.1
cachetools==4.1.0
cchardet==2.1.6
celery==4.1.0
certifi==2020.4.5.1
cffi==1.13.2
cftime==1.1.1.2
chardet==3.0.4
click==7.1.2
cmdstanpy==0.4.0
colorama==0.4.3
convertdate==2.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.16
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.16
django-cors-headers==2.4.0
django-filter==2.0.0
django-import-export==1.2.0
django-rest-swagger==2.1.2
djangorestframework==3.9.0
EasyProcess==0.2.10
elasticsearch==7.6.0
entrypoints==0.3
ephem==3.7.7.1
et-xmlfile==1.0.1
fake-useragent==0.1.11
Flask==1.1.2
flower==0.9.2
geographiclib==1.50
geopy==1.21.0
google-api-core==1.17.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
googletransx==2.4.2
gspread==3.6.0
holidays==0.10.2
httplib2==0.17.3
idna==2.8
importlib-metadata==1.5.0
inflect==4.1.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
itypes==1.1.0
jaraco.itertools==5.0.0
jdcal==1.4
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kiwisolver==1.1.0
kombu==4.1.0
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
lxml==4.4.2
Markdown==3.0.1
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.1.3
MechanicalSoup==0.12.0
mistune==0.8.4
more-itertools==8.2.0
multidict==4.7.5
mysql-connector-python==8.0.18
mysqlclient==1.3.14
nbconvert==5.6.1
nbformat==5.0.4
netCDF4==1.4.2
nltk==3.5
notebook==6.0.3
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
odfpy==1.4.0
olefile==0.46
openapi-codec==1.3.2
openpyxl==2.6.1
packaging==20.3
pandas==1.0.3
pandas-datareader==0.8.1
pandas-market-calendars==1.2
pandas-profiling==1.4.1
pandasql==0.7.3
pandocfilters==1.4.2
parse==1.15.0
parso==0.5.2
patsy==0.5.1
pdblp==0.1.8
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.6.0
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.1
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.19
pyee==7.0.1
Pygments==2.5.2
PyMeeus==0.3.7
PyMySQL==0.9.3
pyodbc==4.0.27
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyppeteer==0.0.25
pyquery==1.4.1
pyrsistent==0.15.7
PySocks==1.7.1
pystan==2.19.1.1
pytest==5.3.5
python-dateutil==2.8.1
pytrends==4.4.0
pytz==2019.3
PyVirtualDisplay==0.2.5
pywin32==227
pywinpty==0.5.7
PyYAML==5.2
pyzmq==18.1.0
qtconsole==4.6.0
regex==2020.4.4
requests==2.22.0
requests-html==0.10.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
schedule==0.6.0
scikit-learn==0.22.1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
setuptools-git==1.2
similarweb-scraper==0.0.3
simplegeneric==0.8.1
simplejson==3.16.0
six==1.13.0
soupsieve==1.8
SQLAlchemy==1.3.12
sqlparse==0.3.1
statsmodels==0.10.1
tablib==0.13.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
tornado==6.0.3
tqdm==4.45.0
traitlets==4.3.3
tweepy==3.8.0
twint==2.1.20
twitter-scraper==0.4.1
uritemplate==3.0.0
urllib3==1.25.7
vine==1.1.4
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xarray==0.15.1
xbbg==0.6.0
xlrd==1.2.0
xlwt==1.3.0
yarl==1.4.2
zipp==2.1.0
